+++ dotnet/Core/Database/Server/Core/Hosting/AllorsServerApplicationBuilderExtensions.cs
@@ -66,6 +66,7 @@ namespace Allors.Server
 
             // app.UseHttpsRedirection();
             app.UseRouting();
+            app.UseRateLimiter();
             app.UseAuthentication();
             app.UseAuthorization();